% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import}
\alias{import}
\alias{importfunctions}
\title{Import Specific User-Defined Functions from Another R Script}
\usage{
import(file = "", only = NULL, except = NULL, envir = .GlobalEnv)
}
\arguments{
\item{file}{Directory of R script that stores all user-defined functions.}

\item{only}{Functions to be imported. Default is NULL.}

\item{except}{Functions that are excluded when importing. Default is NULL.}

\item{envir}{Environment where functions are imported to.}
}
\value{
User-defined functions from another R script
}
\description{
Works similar to import function from Python - user can decide which function(s)
should be included or excluded when importing to the environment from another R script.

It is recommended to execute this function at the start of running current R script.
}
\examples{
\dontrun{
# Suppose there is a .R file (functions.R) that contains functions (foo1, foo2 and foo3):
file.create("functions.R")
writeLines(text = "foo1 <- function(){}\nfoo2 <- function(){}\nfoo3 <- function(){}",
           con = file("functions.R")
)
close(file("functions.R"))

# Import all functions in functions.R
import(file = "functions.R")
ls.str()  # Print what exists in Global Environment
# Import specific functions via functions.R (foo1 and foo3)
import(file = "functions.R", only = c("foo1", "foo3"))
ls.str()  # Print what exists in Global Environment
# Import all functions in functions.R except foo1 and foo3
import(file = "functions.R", except = c("foo1", "foo3"))
ls.str()  # Print what exists in Global Environment
# Running the following code shows an error message as function(s) imported do not exist
import(file = "functions.R", only = c("foo1", "foo03"))
# Error: foo03 could not be found in functions.R

ls.str()  # Print what exists in Global Environment
# Note: despite `foo1` is in functions.R, it is not imported in Global Environment via above command
}

}
\author{
Fendi Tsim (F.Tsim at wbs.ac.uk)
}
