% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{lookup}
\alias{lookup}
\alias{look}
\title{Retrieve names of all customized functions}
\usage{
lookup(file = "")
}
\arguments{
\item{file}{Directory of R script that stores all user-defined functions.}
}
\value{
Names of all user-defined functions from another R script
}
\description{
List names of all customized functions in a R script.
}
\author{
Fendi Tsim (F.Tsim at wbs.ac.uk)
}
